backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
sanpy==0.12.3
pymongo[srv]==3.6
python-dotenv==1.1.1
tweepy==4.16.0
xai-sdk==1.2.0
feedparser==6.0.12
openai==2.6.1